# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Deezer Clone",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/albums/{id}": {
            "get": {
                "tags": [
                    "Albums"
                ],
                "summary": "Obtener detalle de Álbum",
                "operationId": "5743ade2928f8b300e1a72b1ddbf11cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de álbum",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener detalle de álbum exitoso."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/artists/{id}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Obtener detalle de Artista",
                "operationId": "c8cdf7d04a45f87311f8ad46fe5c45ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de artista",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener detalle de artista exitoso."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registrar usuario",
                "operationId": "c7bb21a0d87049e161b8e4c9b0d909f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "country"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nombre",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña",
                                        "type": "string"
                                    },
                                    "country": {
                                        "description": "País",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registrar usuario y obtener sus datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autenticar usuario",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autenticar usuario y obtener sus datos."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cerrar sesión de usuario",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Cerrar sesión de usuario."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/search": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Realizar una búsqueda",
                "operationId": "95baf059062744f8c6fa99abb57d3cf4",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Palabras clave para la búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener resultados de búsqueda exitoso."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}